파일 분리 기능 관리 최적화 sample
