+++ flutter_capston2025/lib/pages/camera_page.dart
@@ -1,0 +1,202 @@
+import 'dart:io';
+import 'dart:ui';
+import 'package:flutter/material.dart';
+import 'package:image_picker/image_picker.dart';
+import 'package:path_provider/path_provider.dart';
+import 'package:path/path.dart' as path;
+
+class CameraPage extends StatefulWidget {
+  final Color themeColor;
+  final VoidCallback onPhotoTaken;
+  const CameraPage({super.key, required this.themeColor, required this.onPhotoTaken});
+
+  @override
+  State<CameraPage> createState() => _CameraPageState();
+}
+
+class _CameraPageState extends State<CameraPage> {
+  File? _lastImage;
+
+  Color _getLighterThemeColor(Color themeColor) {
+    if (themeColor is MaterialColor) {
+      return themeColor.shade100;
+    }
+    return themeColor.withOpacity(0.3);
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    _loadLastImage();
+  }
+
+  Future<void> _loadLastImage() async {
+    final dir = await getApplicationDocumentsDirectory();
+    final photoDir = Directory('${dir.path}/insect_photos');
+    if (await photoDir.exists()) {
+      final files = photoDir
+          .listSync()
+          .whereType<File>()
+          .where((file) => path.basename(file.path).contains("insect_"))
+          .toList();
+      if (files.isNotEmpty) {
+        files.sort((a, b) => b.path.compareTo(a.path));
+        setState(() {
+          _lastImage = files.first;
+        });
+      }
+    }
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+  }
+
+  Future<void> _takePhoto() async {
+    final picker = ImagePicker();
+    final pickedFile = await picker.pickImage(source: ImageSource.camera);
+    if (pickedFile != null) {
+      final dir = await getApplicationDocumentsDirectory();
+      final photoDir = Directory('${dir.path}/insect_photos');
+      if (!await photoDir.exists()) {
+        await photoDir.create(recursive: true);
+      }
+      final fileName = 'insect_${DateTime.now().millisecondsSinceEpoch}.jpg';
+      final newFile = await File(pickedFile.path).copy('${photoDir.path}/$fileName');
+      setState(() {
+        _lastImage = newFile;
+      });
+      widget.onPhotoTaken();
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Stack(
+        children: [
+          Container(
+            decoration: BoxDecoration(
+              image: _lastImage != null
+                  ? DecorationImage(
+                image: FileImage(_lastImage!),
+                fit: BoxFit.cover,
+              )
+                  : null,
+              color: _lastImage == null ? _getLighterThemeColor(widget.themeColor) : null,
+            ),
+            child: _lastImage != null
+                ? BackdropFilter(
+              filter: ImageFilter.blur(sigmaX: 5, sigmaY: 5),
+              child: Container(
+                color: Colors.black.withOpacity(0.2),
+              ),
+            )
+                : null,
+          ),
+          if (_lastImage == null)
+            Center(
+              child: Container(
+                padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 5),
+                decoration: BoxDecoration(
+                  color: widget.themeColor,
+                  borderRadius: BorderRadius.circular(10),
+                ),
+                child: RichText(
+                  text: const TextSpan(
+                    children: [
+                      TextSpan(
+                        text: "사진",
+                        style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 30,
+                          fontWeight: FontWeight.bold,
+                        ),
+                      ),
+                      TextSpan(
+                        text: "이 없습니다!",
+                        style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 20,
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+            ),
+          Positioned(
+            top: 0,
+            left: 0,
+            right: 0,
+            child: Container(
+              height: kBottomNavigationBarHeight,
+              color: widget.themeColor,
+            ),
+          ),
+          if (_lastImage != null)
+            Positioned(
+              top: kBottomNavigationBarHeight + 20,
+              left: 20,
+              right: 20,
+              bottom: 120,
+              child: Container(
+                decoration: BoxDecoration(
+                  border: Border.all(color: widget.themeColor, width: 2),
+                  borderRadius: BorderRadius.circular(10),
+                  boxShadow: [
+                    BoxShadow(
+                      color: Colors.black.withOpacity(0.2),
+                      blurRadius: 5,
+                      offset: const Offset(0, 3),
+                    ),
+                  ],
+                ),
+                child: ClipRRect(
+                  borderRadius: BorderRadius.circular(10),
+                  child: AspectRatio(
+                    aspectRatio: 3 / 2,
+                    child: Image.file(
+                      _lastImage!,
+                      fit: BoxFit.contain,
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          Positioned(
+            bottom: 20,
+            left: 0,
+            right: 0,
+            child: Center(
+              child: ElevatedButton.icon(
+                onPressed: _takePhoto,
+                icon: const Icon(
+                  Icons.camera_alt,
+                  size: 40,
+                  color: Colors.white,
+                ),
+                label: const Text(
+                  '촬영',
+                  style: TextStyle(
+                    color: Colors.white,
+                    fontWeight: FontWeight.bold,
+                    fontSize: 24,
+                  ),
+                ),
+                style: ElevatedButton.styleFrom(
+                  backgroundColor: widget.themeColor,
+                  padding: const EdgeInsets.symmetric(horizontal: 32, vertical: 20),
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(30),
+                  ),
+                ),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
