+++ flutter_capston2025/lib/pages/collection_page.dart
@@ -1,0 +1,156 @@
+import 'dart:io'; // Add this import to use the File class
+import 'package:flutter/material.dart';
+import 'gallery_page.dart';
+import 'dictionary_page.dart';
+
+class CollectionPage extends StatefulWidget {
+  final Color themeColor;
+  final List<File> images; // Now File is recognized
+  final int previewColumns;
+  final VoidCallback onPreviewSetting;
+  final VoidCallback onImageDeleted;
+
+  const CollectionPage({
+    super.key,
+    required this.themeColor,
+    required this.images,
+    required this.previewColumns,
+    required this.onPreviewSetting,
+    required this.onImageDeleted,
+  });
+
+  @override
+  State<CollectionPage> createState() => _CollectionPageState();
+}
+
+class _CollectionPageState extends State<CollectionPage> {
+  Color _getComplementaryColor(Color color) {
+    return Color.fromRGBO(
+      255 - color.red,
+      255 - color.green,
+      255 - color.blue,
+      1.0,
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final complementaryColor = _getComplementaryColor(widget.themeColor);
+
+    return Scaffold(
+      body: Column(
+        children: [
+          Container(height: kBottomNavigationBarHeight, color: widget.themeColor),
+          Expanded(
+            child: Column(
+              children: [
+                Expanded(
+                  child: GestureDetector(
+                    onTap: () => Navigator.push(
+                      context,
+                      MaterialPageRoute(builder: (_) => const DictionaryPage()),
+                    ),
+                    child: AnimatedContainer(
+                      duration: const Duration(milliseconds: 200),
+                      margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+                      decoration: BoxDecoration(
+                        gradient: LinearGradient(
+                          colors: [
+                            complementaryColor,
+                            complementaryColor.withOpacity(0.7),
+                          ],
+                          begin: Alignment.topLeft,
+                          end: Alignment.bottomRight,
+                        ),
+                        borderRadius: BorderRadius.circular(20),
+                        boxShadow: [
+                          BoxShadow(
+                            color: complementaryColor.withOpacity(0.3),
+                            blurRadius: 10,
+                            offset: const Offset(0, 5),
+                          ),
+                        ],
+                      ),
+                      child: const Center(
+                        child: Text(
+                          "도감",
+                          style: TextStyle(
+                            fontSize: 28,
+                            fontWeight: FontWeight.bold,
+                            color: Colors.white,
+                            shadows: [
+                              Shadow(
+                                color: Colors.black26,
+                                blurRadius: 5,
+                                offset: Offset(2, 2),
+                              ),
+                            ],
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+                ),
+                Expanded(
+                  child: GestureDetector(
+                    onTap: () => Navigator.push(
+                      context,
+                      MaterialPageRoute(
+                        builder: (_) => GalleryPage(
+                          themeColor: widget.themeColor,
+                          images: widget.images,
+                          previewColumns: widget.previewColumns,
+                          onPreviewSetting: widget.onPreviewSetting,
+                          onImageDeleted: widget.onImageDeleted,
+                        ),
+                      ),
+                    ),
+                    child: AnimatedContainer(
+                      duration: const Duration(milliseconds: 200),
+                      margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+                      decoration: BoxDecoration(
+                        gradient: LinearGradient(
+                          colors: [
+                            complementaryColor.withOpacity(0.7),
+                            complementaryColor.withOpacity(0.4),
+                          ],
+                          begin: Alignment.topLeft,
+                          end: Alignment.bottomRight,
+                        ),
+                        borderRadius: BorderRadius.circular(20),
+                        boxShadow: [
+                          BoxShadow(
+                            color: complementaryColor.withOpacity(0.3),
+                            blurRadius: 10,
+                            offset: const Offset(0, 5),
+                          ),
+                        ],
+                      ),
+                      child: const Center(
+                        child: Text(
+                          "갤러리",
+                          style: TextStyle(
+                            fontSize: 28,
+                            fontWeight: FontWeight.bold,
+                            color: Colors.white,
+                            shadows: [
+                              Shadow(
+                                color: Colors.black26,
+                                blurRadius: 5,
+                                offset: Offset(2, 2),
+                              ),
+                            ],
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
